+++ magnus/Animation.cpp
@@ -6,7 +6,7 @@ Pather *aiPather;
 Animation::Animation(void)
 {
 }
-Animation::Animation(std::vector<std::vector<int>> tempMap, Ogre::SceneManager* mSceneMgr, Ogre::Camera* mCamera, int SFR, int FFR, int SFD, int FFD, int DFD, int AIS)
+Animation::Animation(std::vector<std::vector<int>> tempMap, Ogre::SceneManager* mSceneMgr, Ogre::Camera* mCamera, int SFR, int FFR, int SFD, int FFD, int DFD, float AIS)
 {
 	mapSizeX = tempMap.size();
 	mapSizeY = tempMap[0].size();
